+++ docs-website/config/navigation.js
@@ -120,8 +120,12 @@ const navigation = [
 				href: '/docs/guides/enable-autocompletion-in-your-ide',
 			},
 			{
-				title: 'Testing',
-				href: '/docs/guides/testing',
+				title: 'Integration testing',
+				href: '/docs/guides/integration-testing',
+			},
+			{
+				title: 'End to end testing',
+				href: '/docs/guides/end-to-end-testing',
 			},
 			{
 				title: 'Debugging',
@@ -264,27 +268,27 @@ const navigation = [
 				href: '/docs/examples/keycloak-openid-connect-authentication',
 			},
 			{
-				title: 'GraphQl Apollo subscriptions',
+				title: 'GraphQL Apollo subscriptions',
 				href: '/docs/examples/graphql-apollo-subscriptions',
 			},
 			{
-				title: 'GraphQl Yoga subscriptions',
+				title: 'GraphQL Yoga subscriptions',
 				href: '/docs/examples/graphql-yoga-subscriptions',
 			},
 			{
-				title: 'GraphQl SSE subscriptions',
+				title: 'GraphQL SSE subscriptions',
 				href: '/docs/examples/graphql-sse-subscriptions',
 			},
 			{
-				title: 'GraphQl WS subscriptions',
+				title: 'GraphQL WS subscriptions',
 				href: '/docs/examples/graphql-ws-subscriptions',
 			},
 			{
-				title: 'GraphQl Hasura WS subscriptions',
+				title: 'GraphQL Hasura WS subscriptions',
 				href: '/docs/examples/graphql-hasura-subscriptions',
 			},
 			{
-				title: 'GraphQl subscriptions hooks',
+				title: 'GraphQL subscriptions hooks',
 				href: '/docs/examples/graphql-subscriptions-hooks',
 			},
 			{