+++ packages/sdk/src/configure/index.ts
@@ -47,9 +47,9 @@ import {
 	OperationExecutionEngine,
 	OperationType,
 	PostResolveTransformationKind,
+	S3UploadProfile as _S3UploadProfile,
 	TypeConfiguration,
 	WebhookConfiguration,
-	S3UploadProfile as _S3UploadProfile,
 	WunderGraphConfiguration,
 } from '@wundergraph/protobuf';
 import { SDK_VERSION } from '../version';
@@ -176,6 +176,11 @@ interface ResolvedDeployment {
 }
 
 export interface S3UploadProfile {
+	/** Whether authentication is required to upload to this profile
+	 *
+	 * @default true
+	 */
+	requireAuthentication?: boolean;
 	/** JSON schema for metadata */
 	meta?: ZodType | object;
 	/**
@@ -228,6 +233,8 @@ interface ResolvedS3UploadConfiguration extends Omit<S3UploadConfiguration, 'upl
 }
 
 export interface ResolvedWunderGraphConfig {
+	// XXX: ResolvedWunderGraphConfig is hashed by several templates.
+	// DO NOT INCLUDE UNSTABLE DATA (paths, times, etc...) in it.
 	application: ResolvedApplication;
 	deployment: ResolvedDeployment;
 	sdkVersion: string;
@@ -259,7 +266,11 @@ export interface ResolvedWunderGraphConfig {
 	serverOptions?: ResolvedServerOptions;
 }
 
-export interface CodeGenerationConfig extends ResolvedWunderGraphConfig {
+export interface CodeGenerationConfig {
+	// Keep ResolvedWunderGraphConfig in a separate field, so it can be
+	// hashed without using any unstable data as the hash input (e.g.
+	// paths like outPath or wunderGraphDir).
+	config: ResolvedWunderGraphConfig;
 	outPath: string;
 	wunderGraphDir: string;
 }
@@ -313,10 +324,15 @@ const resolveConfig = async (config: WunderGraphConfigApplicationConfig): Promis
 		config.apis.push(...graphqlApis);
 	}
 
-	const apps = config.apis;
 	const roles = config.authorization?.roles || ['admin', 'user'];
 
-	const resolved = await resolveApplication(roles, apps, cors, config.s3UploadProvider, config.server?.hooks);
+	const resolved = await resolveApplication(
+		roles,
+		config.apis,
+		cors,
+		config.s3UploadProvider || [],
+		config.server?.hooks
+	);
 
 	const cookieBasedAuthProviders: AuthProvider[] =
 		(config.authentication !== undefined &&
@@ -556,6 +572,7 @@ const resolveUploadConfiguration = (
 			const profileHooks = configurationHooks ? configurationHooks[key] : undefined;
 
 			uploadProfiles[key] = {
+				requireAuthentication: profile.requireAuthentication ?? true,
 				maxAllowedUploadSizeBytes: profile.maxAllowedUploadSizeBytes ?? -1,
 				maxAllowedFiles: profile.maxAllowedFiles ?? -1,
 				allowedMimeTypes: profile.allowedMimeTypes ?? [],
@@ -967,6 +984,7 @@ const ResolvedWunderGraphConfigToJSON = (config: ResolvedWunderGraphConfig): str
 							throw new Error(`error serializing JSON schema for upload profile ${provider.name}/${key}: ${e}`);
 						}
 						uploadProfiles[key] = {
+							requireAuthentication: resolved.requireAuthentication,
 							maxAllowedUploadSizeBytes: resolved.maxAllowedUploadSizeBytes,
 							maxAllowedFiles: resolved.maxAllowedFiles,
 							allowedMimeTypes: resolved.allowedMimeTypes,
@@ -1078,6 +1096,7 @@ const mapDataSource = (source: DataSource): DataSourceConfiguration => {
 		case DataSourceKind.MONGODB:
 		case DataSourceKind.SQLSERVER:
 		case DataSourceKind.SQLITE:
+		case DataSourceKind.PRISMA:
 			const database = source.Custom as DatabaseApiCustom;
 			out.customDatabase = {
 				databaseURL: database.databaseURL,
@@ -1185,10 +1204,12 @@ const loadAndApplyNodeJsOperationOverrides = async (operation: GraphQLOperation)
 
 const applyNodeJsOperationOverrides = (
 	operation: GraphQLOperation,
-	overrides: NodeJSOperation<any, any, any, any, any>
+	overrides: NodeJSOperation<any, any, any, any, any, any, any, any>
 ): GraphQLOperation => {
 	if (overrides.inputSchema) {
-		operation.VariablesSchema = zodToJsonSchema(overrides.inputSchema) as any;
+		const schema = zodToJsonSchema(overrides.inputSchema) as any;
+		operation.VariablesSchema = schema;
+		operation.InternalVariablesSchema = schema;
 	}
 	if (overrides.liveQuery) {
 		operation.LiveQuery = {