+++ scripts/test-examples/all.sh
@@ -20,10 +20,14 @@ set -e
 SKIP="faunadb-nextjs graphql-hasura-subscriptions inject-bearer"
 
 # These are broken
+<<<<<<< HEAD
 SKIP="${SKIP} nextjs-todos vite-swr"
 
 # These are broken on Windows
 SKIP="${SKIP} apollo-federation"
+=======
+SKIP="${SKIP} nextjs-todos"
+>>>>>>> origin/main
 
 # XXX: This breaks only in CI (fastify issue?)
 SKIP="${SKIP} graphql-ws-subscriptions"
@@ -72,7 +76,7 @@ cd ..
 
 if test ${target} = "workspace"; then
     # Restore package.json files in examples
-    find examples -name package.json | grep -v '\.next' | xargs git checkout
+    find examples -maxdepth 2  -name package.json | xargs git checkout
 
     # Restore workspace and lockfile
     git checkout pnpm-workspace.yaml pnpm-lock.yaml